03_multi_cards/pyproject.toml
[project]
name = "03-multi-cards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "uvicorn>=0.34.3",
]
